requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
app==0.0.1
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.30.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
build==1.2.1
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
doctran==0.0.14
docx2txt==0.8
effdet==0.4.1
emoji==2.12.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.0
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
fitz==0.0.1.dev2
Flask==3.0.3
Flask-SSE==1.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
icecream==2.1.3
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.0.0
importlib_resources==6.4.0
interchange==2021.0.4
iopath==0.1.10
ipython==8.26.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json_repair==0.28.4
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.16
langchain-chroma==0.1.3
langchain-community==0.2.16
langchain-core==0.2.38
langchain-experimental==0.0.64
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langcodes==3.4.0
langdetect==1.0.9
langgraph==0.2.14
langgraph-checkpoint==1.0.6
langsmith==0.1.98
language_data==1.2.0
layoutparser==0.3.4
lazy_loader==0.4
lmdb==1.5.1
looseversion==1.3.0
lxml==4.9.4
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
neo4j==5.23.1
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.3.1
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.18.1
openai==0.27.10
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
pansi==2020.7.3
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.3
pexpect==4.9.0
phonenumbers==8.13.42
pikepdf==9.1.1
pillow==10.4.0
pillow_heif==0.18.0
portalocker==2.10.1
posthog==3.5.0
preshed==3.0.9
presidio_analyzer==2.2.355
presidio_anonymizer==2.2.355
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
prov==2.0.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py2neo==2021.2.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.17
pydantic_core==2.20.1
pydot==3.0.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
pypdf==4.3.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==1.0.2
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.2
rapidfuzz==3.9.6
rapidocr-onnxruntime==1.3.24
rdflib==6.3.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.32
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
thinc==8.2.5
tifffile==2024.7.24
tiktoken==0.5.2
timm==1.0.8
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
traitlets==5.14.3
traits==6.3.2
transformers==4.44.0
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.1
unstructured-client==0.25.4
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
urllib3==2.2.2
uvicorn==0.30.5
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.23.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.19.2
